requirements.txt
# requirements.txt - Python dependencies for the news aggregator application
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.1.6
requests>=2.31.0
python-dotenv==1.2.2
nltk>=3.9.3
transformers==4.35.2
torch
newspaper3k>=0.2.8
google-generativeai